+++ frontend/react-app/src/dijkstra/dijkstra.js
@@ -1,6 +1,8 @@
 import { aperture, assoc, pipe, range, splitEvery, update } from 'ramda';
+import * as vis from 'vis';
 
-export const parseDijkstraResponse = (response, setResult, setControl) => {
+export const dijkstraParser = (response, setControl, setResult) => {
+
   if (response.data.status === 'OK') {
 
     // グラフの素
@@ -104,3 +106,42 @@ export const parseDijkstraResponse = (response, setResult, setControl) => {
     setControl(3)
   }
 }
+
+export const drawDijkstraGraph = (id, graphs, currentStep) => {
+  const container = document.getElementById(id);
+  const data = graphs[currentStep];
+  const options = {
+    edges: {
+      chosen: false,
+      color: {
+        color: '#848484'
+      }
+    },
+    
+    // todo: ここをうまいこといじると、リンクの重複を解消できそう
+    // 要ドキュメントよみこみ https://visjs.github.io/vis-network/docs/network/
+    // physics: {
+    //   enabled: true,
+    //   hierarchicalRepulsion: {
+    //     centralGravity: 0.0,
+    //     springLength: 500,
+    //     springConstant: 0.01,
+    //     nodeDistance: 100,
+    //     damping: 0.09
+    //   },
+    //   solver: 'hierarchicalRepulsion'
+    // },
+    interaction: {
+      dragView: false,
+      zoomView: false
+    },
+    layout: {
+      randomSeed: 0
+    },
+    nodes: {
+      chosen: false
+    },
+  };
+  new vis.Network(container, data, options);
+}
+
